+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 20.10+ (in development)
 ------------------------------------------------------------------------
+- Feature: [#690] Automatically save the game at regular intervals.
 - Fix: [#151] Mouse moves out of window when looking around.
 - Fix: [#588] 'Cancel or Show Last Announcement' shortcut doesn't close announcements.
 - Fix: [#679] Crash when changing ground texture.
+- Change: [#690] Default saved game directory is now in OpenLoco user directory.
 
 20.10 (2020-10-25)
 ------------------------------------------------------------------------